ref: slurm connector 1/n (#3476)
* ref: slurm connector 1/n

* ref: slurm connector 1/n

* ref: slurm connector 1/n

* ref: slurm connector 1/n
