+++ pytorch_lightning/core/lightning.py
@@ -862,7 +862,7 @@ class LightningModule(ABC, DeviceDtypeModuleMixin, GradInformation, ModelIO, Mod
         except Exception:
             root_node = '127.0.0.1'
 
-        root_node = self.trainer.accelerator_connector.resolve_root_node_address(root_node)
+        root_node = self.trainer.slurm_connector.resolve_root_node_address(root_node)
         os.environ['MASTER_ADDR'] = root_node
 
     def init_ddp_connection(self, global_rank: int, world_size: int, is_slurm_managing_tasks: bool = True) -> None:
