+++ tests/models/test_amp.py
@@ -155,10 +155,10 @@ def test_amp_gpu_ddp_slurm_managed(tmpdir):
     assert result == 1, 'amp + ddp model failed to complete'
 
     # test root model address
-    assert trainer.accelerator_connector.resolve_root_node_address('abc') == 'abc'
-    assert trainer.accelerator_connector.resolve_root_node_address('abc[23]') == 'abc23'
-    assert trainer.accelerator_connector.resolve_root_node_address('abc[23-24]') == 'abc23'
-    assert trainer.accelerator_connector.resolve_root_node_address('abc[23-24, 45-40, 40]') == 'abc23'
+    assert trainer.slurm_connector.resolve_root_node_address('abc') == 'abc'
+    assert trainer.slurm_connector.resolve_root_node_address('abc[23]') == 'abc23'
+    assert trainer.slurm_connector.resolve_root_node_address('abc[23-24]') == 'abc23'
+    assert trainer.slurm_connector.resolve_root_node_address('abc[23-24, 45-40, 40]') == 'abc23'
 
 
 def test_cpu_model_with_amp(tmpdir):
